ROX-28353: use NVD 2.0 JSON feeds
